% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_chr_signal.R
\name{average_chr_signal}
\alias{average_chr_signal}
\title{Average signal genome-wide and on each chromosome}
\usage{
average_chr_signal(
  gr,
  remove_cen = FALSE,
  genome,
  cen_region_length = 50000,
  mean_norm = FALSE,
  order_chrs = FALSE
)
}
\arguments{
\item{gr}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{remove_cen}{Logical indicating whether to remove regions around
centromeres. Defaults to \code{FALSE}.}

\item{genome}{Character string indicating reference genome used to align
the data. Must be provided when \code{remove_cen = TRUE}, in order to load
appropriate centromere data, or \code{order_chrs = TRUE}, in order to load
chromosome length data. Accepts one of the following strings:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}

\item{cen_region_length}{Integer indicating the length (in bp) of the region
to remove (centered on the centromere of each chromosome).
Defaults to 50'000 bp.}

\item{mean_norm}{Logical indicating whether to normalize by average
genome-wide score. This adds two columns to the dataframe (first element of
the output list): mean ratio (\code{avrg_signal_mean_ratio}) and
mean-subtracted (\code{avrg_signal_mean_subtracted}).
Defaults to \code{FALSE}.}

\item{order_chrs}{Logical indicating whether to order rows by chromosome
length (in increasing order). Defaults to \code{FALSE}.}
}
\value{
List with two elements:
\enumerate{
  \item \code{seq_avrg} Average \code{score} on each sequence (dataframe)
  \item \code{genome_avrg} Average \code{score} genome-wide (named vector)
}
}
\description{
Computes average signal (in the \code{score} \code{GRanges} metadata column)
genome-wide and on each chromosome (each individual sequence determined by
\code{seqnames}).
}
\examples{
\dontrun{
average_chr_signal(GRanges_object)

average_chr_signal(anti_Rec8, remove_cen=TRUE, mean_norm=TRUE, order_chrs=TRUE)

average_chr_signal(anti_Rec8, remove_cen=TRUE, genome = 'SK1Yue',
                   cen_region_length=40000, mean_norm=TRUE, order_chrs=TRUE)
}
}
