% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_signal.R
\name{average_signal}
\alias{average_signal}
\title{Average signal genome-wide and on each chromosome}
\usage{
average_signal(gr)
}
\arguments{
\item{gr}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}
}
\value{
List of two named numeric vectors with average \code{score} on:
\enumerate{
  \item \code{seq_avrg} Each individual sequence
  \item \code{genome_avrg} Whole genome
}
}
\description{
Computes average signal (in the \code{score} \code{GRanges} metadata column)
genome-wide and on each chromosome (each individual sequence determined by
\code{seqnames}).
}
\examples{
\dontrun{
average_signal(GRanges_objct)

average_signal(WT)
}
}
