% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_signal_track.R
\name{compress_signal_track}
\alias{compress_signal_track}
\title{Average signal variable over specified position windows}
\usage{
compress_signal_track(df, window_size)
}
\arguments{
\item{df}{Input signal track data as a \code{data.frame} with at least two
columns:
\enumerate{
  \item Numeric column with positions to partition
  \item Numeric column with values to average in each partition
}
No default.}

\item{window_size}{Integer indicating the length (in bp) of the window (or
partition) to average signal in. No default.}
}
\description{
Partitions the position variable (first column of input data frame) in
windows of the size specified in \code{window_size} and computes the average
signal (second column) in each partition. Most common use case is to compress
the data for plotting.
}
\examples{
\dontrun{
compress_signal_track(df=signal_data, window_size=10)

compress_signal_track(chrI_signal, 100)
}
}
