% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_density_plot2.R
\name{facs_density_plot2}
\alias{facs_density_plot2}
\title{FACS density plot}
\usage{
facs_density_plot2(dir, strain_code, gate, plot_color = "black",
  plot_transparency = 0.8, file_format = "jpeg", user_input = TRUE)
}
\arguments{
\item{dir}{Path to directory containing \code{.fcs} files. No default.}

\item{strain_code}{The lab's database numeric code identifier of the yeast
strain. No default.}

\item{gate}{Optional two-element vector containing gating limits. No default
(corresponds to no gating).}

\item{plot_color}{Optional R color to fill the plot. Defaults to
\code{black}.}

\item{plot_transparency}{Floating point number between 0 and 1 indicating 
transparency of the plot (corresponds to the plotting \code{alpha} argument:
0 is full transparency and 1 is completely opaque). Defaults to \code{0.8}.
(corresponds to no gating).}

\item{file_format}{Character string indicating the plot file format. Must be
one of the options \code{jpeg} and \code{pdf}. Defaults to \code{jpeg}.}

\item{user_input}{Logical indicating whether to ask user before saving plot
to file. Defaults to \code{TRUE}.}
}
\value{
Density plot file in the same directory as the \code{.fcs} files.
}
\description{
Given a path to a directory containing \code{.fcs} files and a pattern to
match in the file names, generates density plots of FACS data across time
points.
}
\section{Details}{

The function is designed to search all \code{.fcs} files (named according to
the convention "Well strain_timepoint.fcs". e.g. "A01 119_0.fcs") for those
matching the provided pattern (yeast strain code) and build a density plot.
An ungated version of the plot is printed to the screen and the final plot is
saved to a file in the selected format in the same directory as the
\code{.fcs} files.
}

\examples{
\dontrun{
facs_density_plot2(119)

facs_density_plot2(dir='~/Desktop', strain_code=119,
                   gate=c(1500000, 7000000), file_format='pdf')

facs_density_plot2(dir='~/Desktop', strain_code=119,
                   gate=c(1500000, 7000000), plot_color='orange',
                   plot_transparency=1)
}
}
