% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dsb_hotspots.R
\name{get_dsb_hotspots}
\alias{get_dsb_hotspots}
\title{Get DSB hotspots}
\usage{
get_dsb_hotspots(genome, as_df = FALSE)
}
\arguments{
\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
}
No default.}

\item{as_df}{Logical specifying whether the output should be returned as a
\code{data frame}. If \code{FALSE}, output is a \code{GRanges} object.
Defaults to \code{FALSE}.}
}
\value{
Either a \code{Granges} or a {data.frame} object containing DSB
hotspot start and end coordinates, as well as intensity.
}
\description{
Returns double-strand break (DSB) hotspots mapped using the Spo11 oligo
technique. FASTQ data aligned from scratch for the package. Hotspots
originally published in
\href{https://www.ncbi.nlm.nih.gov/pubmed/21376234}{Pan \emph{et al.} 2011}
and in
\href{https://www.ncbi.nlm.nih.gov/pubmed/24717437}{Thacker \emph{et al.} 
2014}).
}
\examples{
\dontrun{
get_dsb_hotspots(genome='SK1Yue')

get_dsb_hotspots(genome='sacCer3', as_df=TRUE)
}
}
