% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gff.R
\name{get_gff}
\alias{get_gff}
\title{Get GFF}
\usage{
get_gff(genome)
}
\arguments{
\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}
}
\value{
GFF table as \code{GRanges} object.
}
\description{
Returns GFF table for specified genome.
}
\examples{
\dontrun{
get_gff(genome='SK1Yue')

get_gff(genome='sacCer3')
}
}
