% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intergenic_regions.R
\name{get_intergenic_regions}
\alias{get_intergenic_regions}
\title{Get intergenic regions}
\usage{
get_intergenic_regions(genome, as_gr = FALSE)
}
\arguments{
\item{genome}{Character object specifying the reference genome; accepts one
of the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}

\item{as_gr}{Logical specifying whether the output should be returned as a
\code{GRanges} object. If \code{FALSE}, output is a \code{data.frame}. Can
only be used if \code{genome="SK1Yue"}, because \code{"SK1"} and
\code{"sacCer3"} contain overlapping genes, resulting in negative intervals
(not allowed by \code{IRanges}). Defaults to \code{FALSE}.}
}
\value{
R data frame containing coordinates and annotations for all
intergenic regions in the genome.
}
\description{
Returns table of intergenic regions for specified genome.
}
\examples{
\dontrun{
get_intergenic_regions(genome='SK1Yue')

get_intergenic_regions(genome='sacCer3')

}
}
