% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Red1_summits.R
\name{get_Red1_summits}
\alias{get_Red1_summits}
\title{Get Red1 ChIP-seq summits in WT}
\usage{
get_Red1_summits(genome, as_df = FALSE)
}
\arguments{
\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
}
No default.}

\item{as_df}{Logical specifying whether the output should be returned as a
\code{data frame}. If \code{FALSE}, output is a \code{GRanges} object.
Defaults to \code{FALSE}.}
}
\value{
Either a \code{Granges} or a {data.frame} object containing Red1
summits.
}
\description{
Returns all Red1 ChIP-seq summits called in the average of four WT replicates
for which -log10 of the peak's p-value is equal or greater than 20.
}
\examples{
\dontrun{
get_Red1_summits(genome='SK1Yue')

get_Red1_summits(genome='sacCer3', as_df=TRUE)
}
}
