% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wiggle.R
\name{import_wiggle}
\alias{import_wiggle}
\title{Import wiggle data generated by the lab's ChIP-seq analysis pipeline}
\usage{
import_wiggle(path, local_copy = TRUE)
}
\arguments{
\item{path}{Path to the folder containing the wiggle data. No default.}

\item{local_copy}{Logical indicating whether to create local copy of target
file before importing data. If \code{TRUE} a local copy is automatically
created and deleted after use. Use this argument to avoid reading files
directly from shared locations (namely \code{LabShare}).
Defaults to \code{TRUE}.}
}
\value{
\code{GRanges} object.
}
\description{
Given a path to a collection of 16 tab-separated wiggle data files (one per
yeast chromosome), imports data into a \code{GRanges} object.
}
\examples{
\dontrun{
x <- import_wiggle('wiggle_data/')

dot1 <- import_wiggle('/Volumes/LabShare/HTGenomics/HiSeqOutputs/\
                      AveReps_SacCer3_MACS2/Red1-dot1D-195-16-reps-MACS2/',
                      local_copy=TRUE)
}
}
