% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opening_act2.R
\name{opening_act2}
\alias{opening_act2}
\title{Standard analysis of ChIP-seq experiments}
\usage{
opening_act2(signal_data, genome, genotype, chip_target, sample_id,
  output_path = "/Volumes/LabShare/HTGenomics/Opening_act/",
  user_input = TRUE, run_chr_size_bias = TRUE, run_centromeres = TRUE,
  run_rDNA = TRUE, run_telomeres = TRUE, run_dsb_hotspots = TRUE,
  run_axis = TRUE, run_meta_orf = TRUE)
}
\arguments{
\item{signal_data}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
}
No default.}

\item{genotype}{Character object indicating the relevant strain mutations.
Just use \code{genotype=WT}, for example, if there are no relevant mutations.
No default.}

\item{chip_target}{Character object indicating the ChIP target protein.
No default.}

\item{sample_id}{Character object indicating the sample ID, including the ID
used in the analysis pipeline (with a date) and the read mapping conditions
(see examples below). If \code{user_input=TRUE} the function asks the user to
check that the provided \code{sample_id} matches the required format before
proceeding with the analysis. No default.}

\item{user_input}{Logical indicating whether to ask user to check the format
of the \code{sample_id} argument. Defaults to \code{TRUE}.}

\item{run_chr_size_bias}{Logical indicating whether to run the chromosome
size bias analysis. All included analysis can be turned off in this way.
Defaults to \code{TRUE}.}

\item{run_centromeres}{Logical indicating whether to run the signal at 
centromere analysis. Defaults to \code{TRUE}.}

\item{run_rDNA}{Logical indicating whether to run the signal flanking rDNA
analysis. Defaults to \code{TRUE}.}

\item{run_telomeres}{Logical indicating whether to run the signal at
sub-telomeric region analysis. Defaults to \code{TRUE}.}

\item{run_dsb_hotspots}{Logical indicating whether to run the signal at DSB
hotspot analysis. Defaults to \code{TRUE}.}

\item{run_axis}{Logical indicating whether to run the signal at axis binding site
analysis. Defaults to \code{TRUE}.}

\item{run_meta_orf}{Logical indicating whether to run the meta ORF analysis.
Defaults to \code{TRUE}.}

\item{ouput_path}{Character object with a valid path to directory to save
output files at.
Defaults to \code{ouput_path='/Volumes/LabShare/HTGenomics/Opening_act/'}}
}
\value{
A new folder at the specified location (default is
\code{.../LabShare/HTGenomics/Opening_act/}) containing output plots (as .pdf
files) of the included of the following analyses:
\enumerate{
  \item Chromosome size bias
  \item Signal at centromeres
  \item Signal flanking rDNA
  \item Signal at sub-telomeric regions
  \item Signal at DSB hotspots
  \item Signal at axis binding sites
  \item Signal at meta ORF
}
}
\description{
Runs the lab's standard analysis of ChIP-seq experiment data and produces
several .pdf files of analysis plots to be kept in a new folder at
\code{.../LabShare/HTGenomics/Opening_act/}.
}
\examples{
\dontrun{
opening_act2(signal_data=WT, genome="sacCer3", genotype="WT",
             chip_target="Red1", sample_id="AH119C-040114-sacCer3-2mis")

opening_act2(set1_wiggle, genome="sacCer3", "set1", "Red1",
             "AH8584b-16032016-sacCer3-2mis")

opening_act2(rec8, genome="SK1Yue", "rec8", "Red1",
             "AH8115b-24042015-SK1Yue-PM", user_input=FALSE,
             run_meta_orf=FALSE)
            
opening_act2(signal_data=WT, genome="sacCer3", genotype="WT",
             chip_target="Red1", sample_id="AH119C-040114-sacCer3-2mis",
             run_meta_orf=FALSE)

opening_act2(signal_data=WT, genome="sacCer3", genotype="WT",
             chip_target="Red1", sample_id="AH119C-040114-sacCer3-2mis",
             output_path='~/Desktop'
             run_chr_size_bias=FALSE, run_centromeres=FALSE, run_rDNA=FALSE,
             run_telomeres=FALSE, run_dsb_hotspots=TRUE, run_axis=TRUE,
             run_meta_orf=FALSE)
}
}
