% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opening_act2_batch_run.R
\name{opening_act2_batch_run}
\alias{opening_act2_batch_run}
\title{Batch run of \code{\link{opening_act2}} function}
\usage{
opening_act2_batch_run(
  input_data_file,
  ref_genome,
  output_path = "/Volumes/LabShare/HTGenomics/Opening_act/"
)
}
\arguments{
\item{input_data_file}{String indicating path to file containing
instructions. The file must be in CSV format with the following columns
(column names in the file must be reproduced exactly as below):
\enumerate{
  \item \strong{path} Path to bedGraph file.
  \item \strong{genotype} String indicating the relevant strain
  mutations, used as the namesake argument to \code{\link{opening_act2}}.
  \item \strong{chip_target} String indicating the ChIP target protein, to be
  used as the namesake argument to \code{\link{opening_act2}}.
  \item \strong{sample_id} The sample ID, to be used as the namesake argument
  to \code{\link{opening_act2}}.
}
No default.}

\item{ref_genome}{Character object specifying the genome version, to be used
as the \code{genome} argument to \code{\link{opening_act2}}; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
}
No default.}

\item{output_path}{Character object with a valid path to directory to save
output files at. Defaults to
\code{'/Volumes/LabShare/HTGenomics/Opening_act/'}.}
}
\description{
This function will take a CSV file containing the path to multiple data
folders and argument labels and run \code{\link{opening_act2}} function for
each data set in the file. It allows you to run a batch of data sets
automatically.
}
\examples{
\dontrun{
opening_act_batch_run(input_data_file="~/Desktop/inputs.csv",
                      ref_genome="SK1Yue")

opening_act_batch_run(input_data_file="~/Desktop/inputs.csv",
                      ref_genome="sacCer3", output_path='~/Desktop/')
}
}
