% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_position.R
\name{signal_at_position}
\alias{signal_at_position}
\title{Signal around provided genomic positions}
\usage{
signal_at_position(signal_data, positions, position_names, up_ext = 1000,
  down_ext = 1000, write_to_file = FALSE)
}
\arguments{
\item{signal_data}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{positions}{Position data as a \code{GRanges} object. Will also accept
an interval \code{GRanges} object of constant width (extended by specified
\code{up_ext} and \code{down_ext}). No default.}

\item{position_names}{String indicating the column containing the position
names: either the \code{GRanges} 'seqnames' or a metadata column. Will be
added to the output to identify the different positions. No default.}

\item{up_ext}{Integer indicating number of bp to collect upstream of start
position. Defaults to \code{1000}.}

\item{down_ext}{Integer indicating number of bp to collect downstream of end
position. Deafaults to \code{1000}.}

\item{write_to_file}{Logical indicating whether output should be written to a
.txt file (in current working directory). If \code{to_file = FALSE} the
function returns output R object. Defaults to \code{FALSE}.}
}
\value{
A data frame with 1001 columns:
\enumerate{
  \item \code{gene} First column with the gene IDs (if found in input gff)
  \item \code{t1 to t1000} 1000 columns with signal at each meta position.
}
Columns t1 to t1000 correspond to the following meta positions:
\enumerate{
  \item \code{t1 to t250} Upstream flank (1/2 of gene length)
  \item \code{t251 to t750} ORF
  \item \code{t751 to t1000} Downstream flank (1/2 of gene length)
}
}
\description{
Pulls out signal in intervals of the defined length centered at provided
positions.\cr
\cr \cr
\strong{Note:} Our ChIP-seq data always contains gaps with missing data. The
affected intervals will contain "NA" values in the output. As a warning, the
number of affected genes is printed to the console.\cr
}
\examples{
\dontrun{
signal_at_position(WT, Red1_summits, 'name')

signal_at_position(WT, centromeres, 'seqnames',
                   up_ext = 5000, down_ext = 5000,
                   write_to_file = TRUE)
}
}
