% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_flanking_rDNA.R
\name{signal_flanking_rDNA}
\alias{signal_flanking_rDNA}
\title{Signal flanking rDNA}
\usage{
signal_flanking_rDNA(signal_data, flank_length = 40000, genome)
}
\arguments{
\item{signal_data}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{flank_length}{Integer specifying the length (in bp) of the windows to
collect signal for up and downstream of the rDNA. Defaults to 40000 (i.e.
40 kb).}

\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}
}
\description{
Pulls out ChIP signal at specified window sizes flanking the rDNA region on
chromosome 12.
}
\examples{
\dontrun{
signal_flanking_rDNA(WT, genome = 'SK1Yue')

signal_flanking_rDNA(WT, flank_length=50000, genome = 'sacCer3')
}
}
