% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_from_telomeres2.R
\name{signal_from_telomeres2}
\alias{signal_from_telomeres2}
\title{Collect signal from telomeres for all chromosome arms}
\usage{
signal_from_telomeres2(signal_data, length_to_collect = 1e+05,
  averaging_window = 1, genome)
}
\arguments{
\item{signal_data}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{length_to_collect}{Integer specifying the length (in bp) of the region
to collect signal for, starting form the telomeres. Defaults to 100000 (i.e.
100 kb).}

\item{averaging_window}{Integer specifying the length (in bp) of the window
to average the signal in.
Defaults to 1 (i.e. 1 bp: no window averaging).}

\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}
}
\value{
An R data frame containing 32 rows (one for each chromosome arm) and
a number of columns dependent on the specified \code{length_to_collect}. The
following columns are included:
\enumerate{
  \item \code{chr} The chromosome number
  \item \code{arm} The chromosome arm; one of "L" and "R"
  \item \code{size-cat} The chromosome size category; one of "small" and
  "large"
  \item \code{t1:tn} Signal columns for each position; number depends on the
  the specified collection length (\code{n = length_to_collect})
}
}
\description{
Pulls out the ChIP signal starting at chromosome ends inward up to a
specified distance.
}
\section{Warning}{

The fact that some sub-telomeric sequences may be incomplete
in the genome means that in some cases we have ChIP-seq data mapping all the
way to the very end of the available sequence. This, together with the read
extension performed by MACS, leads to some cases where the last positions in
the signal data for the right chromosome arms are higher than the annotated
length of the respective chromosome. As a result, some negative position
values may appear in the output of this function (typically not off by more
than 150 bp).
}

\examples{
\dontrun{
signal_from_telomeres2(WT, genome = "SK1Yue")

signal_from_telomeres2(WT, length_to_collect = 50000, averaging_window = 100,
                       genome = "sacCer3")
}
}
