% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_mean_and_ci.R
\name{signal_mean_and_ci}
\alias{signal_mean_and_ci}
\title{Mean, SD, SEM, and confidence interval}
\usage{
signal_mean_and_ci(signal_data, ci = 0.95, rep_bootstrap = 1000, na_rm = TRUE)
}
\arguments{
\item{signal_data}{Input signal data matrix (or an object that can be coerced
to class \code{matrix}). No default.}

\item{ci}{Numeric value between 0 and 1 representing the confidence interval
percentage. Defaults to \code{0.95}.}

\item{rep_bootstrap}{Integer specifying the number of times to repeat the
bootstrapping procedure. Defaults to \code{1000}.}

\item{na_rm}{Logical indicating whether to remove \code{NA} values in each
column before calculating the mean and bootstrapping. Defaults to \code{TRUE}.}
}
\value{
A matrix with as many rows as there are columns in the input and
three columns:
\enumerate{
  \item \code{Mean} Input column mean
  \item \code{Lower} Lower limit of the bootstrapping confidence interval
  \item \code{Upper} Upper limit of the bootstrapping confidence interval
  \item \code{SD} Standard deviation
  \item \code{SEM} Standard error of the mean
}
}
\description{
Calculates the mean value and confidence interval, as well as standard
deviation and standard error of the mean, for every column in an input table.
Confidence intervals are calculated using the "Empirical bootstrap" technique
(based on random sampling with replacement). The input will typically be a
table of signal values for different obervations (across rows) over a set of
positions (across columns).
}
\examples{
\dontrun{
signal_mean_and_ci(WT_signal_at_orf)

signal_mean_and_ci(WT_signal_at_summits, ci=0.90)

signal_mean_and_ci(WT_signal_at_dsb, bootstrap_reps=5000)
}
}
