% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_track_correlation.R
\name{signal_track_correlation}
\alias{signal_track_correlation}
\title{Compute correlation between signal track data for two samples}
\usage{
signal_track_correlation(signal_data_A, signal_data_B, genome,
  method = "pearson")
}
\arguments{
\item{signal_data_A}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{signal_data_B}{Second samples' signal track data in the same format as
\code{signal_data_A} object. No default.}

\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}

\item{method}{Character object specifying which correlation coefficient is to
be computed. Will be the input to the \code{method} argument of function
\code{cor} of the \code{stats} package. Accepts one of (can be abbreviated):
\enumerate{
  \item \code{"pearson"}
  \item \code{"kendall"}
  \item \code{"spearman"}
}
Defaults to \code{"pearson"}.}
}
\value{
A floating point value corresponding to the signal correlation.
}
\description{
Returns correlation between ChIP-seq signal of two samples. Useful to check
replicates or compare different samples.
}
\examples{
\dontrun{
signal_track_correlation(WT, dot1, genome = "SK1Yue")

signal_track_correlation(WT_A, WT_B, genome = "sacCer3", method="spearman")
}
}
