% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikein_normalization_factor.R
\name{spikein_normalization_factor}
\alias{spikein_normalization_factor}
\title{Compute spike-in normalization factor}
\usage{
spikein_normalization_factor(ref_chip_counts, ref_input_counts,
  test_chip_counts, test_input_counts)
}
\arguments{
\item{ref_chip_counts}{Path to reference ChIP samples' read counts file.
No default.}

\item{ref_input_counts}{Path to reference input samples' read counts file.
No default.}

\item{test_chip_counts}{Path to test ChIP samples' read counts file.
No default.}

\item{test_input_counts}{Path to test input samples' read counts file.
No default.}
}
\value{
Numeric normalization factor.
}
\description{
Computes spike-in normalization factor between two spiked-in samples. Inputs
text files containing counts of aligned reads per chromosome of a hybrid
SK1:S288C genome.
}
\examples{
\dontrun{
spikein_normalization_factor(ref_chip_counts='Counts_AH119_chip.txt',
                             ref_input_counts='Counts_AH119_input.txt',
                             test_chip_counts='Counts_AH8104_chip.txt',
                             test_input_counts='Counts_AH8104_input.txt')
}
}
