% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed_time.R
\name{elapsed_time}
\alias{elapsed_time}
\title{Return elapsed time}
\usage{
elapsed_time(t0, t1)
}
\arguments{
\item{t0}{Numeric start time representing seconds. No default.}

\item{t1}{Numeric end time representing continuous second count from
\code{t0}. No default.}
}
\value{
\code{String} containing elapsed time.
}
\description{
Given a start and end time in seconds, returns elapsed time in sec., min. or
hrs, as appropriate.
}
\examples{
\dontrun{
start <- proc.time()[3]
end <- proc.time()[3]
elapsed_time(t0=start, t1=end)

start <- proc.time()
elapsed_time(start[3], proc.time()[3])
}
}
