% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chr_coordinates.R
\name{get_chr_coordinates}
\alias{get_chr_coordinates}
\title{Get chromosome coordinates}
\usage{
get_chr_coordinates(genome = "SK1Yue", as_df = FALSE)
}
\arguments{
\item{genome}{Character object specifying the genome version; accepts one of
the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
Default is \code{"SK1Yue"}.}

\item{as_df}{Logical specifying whether the output should be returned as a
\code{data frame}. If \code{FALSE}, output is a \code{GRanges} object.
Defaults to \code{FALSE}.}
}
\value{
Centromere start and end, as well as chroosome lengths, for each
chromosome.
}
\description{
Returns chromosome length and centromere coordinates for specified genome.
}
\examples{
\dontrun{
get_chr_coordinates()

get_chr_coordinates(genome='sacCer3', as_df=FALSE)

get_chr_coordinates(genome='SK1', as_df=TRUE)
}
}
