% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intergenic_regions.R
\name{get_intergenic_regions}
\alias{get_intergenic_regions}
\title{Get intergenic regions}
\usage{
get_intergenic_regions(genome)
}
\arguments{
\item{genome}{Character object specifying the reference genome; accepts one
of the following options:
\enumerate{
  \item \code{"SK1Yue"}
  \item \code{"sacCer3"}
  \item \code{"SK1"}
}
No default.}
}
\value{
R data frame containing coordinates and annotations for all
intergenic regions in the genome.
}
\description{
Returns table of intergenic regions for specified genome.
}
\examples{
\dontrun{
get_intergenic_regions(genome='SK1Yue')

get_intergenic_regions(genome='sacCer3')

}
}
