% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_orf2.R
\name{signal_at_orf2}
\alias{signal_at_orf2}
\title{Signal at all ORFs genome-wide normalized to constant length (meta ORF)}
\usage{
signal_at_orf2(
  signal_data,
  gff,
  window_size = 1,
  write_to_file = FALSE,
  file_name
)
}
\arguments{
\item{signal_data}{Input signal track data as a \code{GRanges} object (see
\code{?"GRanges-class"} for more details). To load wiggle and bedGraph data
run \code{\link{import_wiggle}} and \code{\link{import_bedGraph}},
respectively. No default.}

\item{gff}{Either a path to a gff file or loaded gff data as a \code{GRanges}
object. No default.}

\item{window_size}{Integer specifying the width in bp of a window to average
signal in. Defaults to \code{1} (no averaging).}

\item{write_to_file}{Logical indicating whether output should be written to a
.txt file. The file will be saved at location provided in \code{file_name}
argument or at the current working directory if only a file name is provided).
If \code{write_to_file = FALSE} the function returns output R object.
Defaults to \code{FALSE}.}

\item{file_name}{Character string indicating file name when writing output to
a .txt file. Must be provided if \code{write_to_file = TRUE}. No default.}
}
\value{
An \code{EnrichedHeatmap} matrix with 1000 columns containing signal
at each meta position. Columns correspond to the following meta positions:
\enumerate{
  \item \code{t1 to t250} Upstream flank (1/2 of gene length)
  \item \code{t251 to t750} ORF
  \item \code{t751 to t1000} Downstream flank (1/2 of gene length)
}
}
\description{
Pulls out the ChIP signal over all ORFs in the genome normalized to a
constant length. Collects the signal over each ORF plus both flanking regions
(1/2 the length of the ORF on each side) and scales them all to the same
value (1000). This means that for two example genes with lengths of 500 bp
and 2 kb, flanking regions of 250 bp and 1 kb will be collected up and
downstream, respectively. The whole region is then rescaled to a length of
1000, corresponding to a gene length of 500 plus 250 for each flanking
region.
}
\section{Warning}{

Our ChIP-seq data always contains gaps with missing data. The
affected genes will contain "NA" values in the output. As a warning, the
number of affected genes is printed to the console.
}

\examples{
\dontrun{
signal_at_orf2(WT, gff = gff)

signal_at_orf2(WT, gff = "S288C_annotation_modified.gff",
               write_to_file = TRUE, file_name = "WT_S288C_metaORF.txt")
}
}
